# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/kyc/bvn": {
      "post": {
        "description": "Submit a BVN for identity verification. One per user across all businesses.",
        "operationId": "KycController_verifyBvn",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyBvnDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "BVN verification submitted successfully."
          },
          "400": {
            "description": "Validation failed or the request body/params were malformed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired authentication.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource could not be found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "The resource already exists or is in a conflicting state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Verify BVN",
        "tags": [
          "KYC Verification"
        ]
      }
    },
    "/api/v1/kyc/nin": {
      "post": {
        "description": "Submit a NIN for identity verification. One per user across all businesses.",
        "operationId": "KycController_verifyNin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyNinDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "NIN verification submitted successfully."
          },
          "400": {
            "description": "Validation failed or the request body/params were malformed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired authentication.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource could not be found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "The resource already exists or is in a conflicting state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Verify NIN",
        "tags": [
          "KYC Verification"
        ]
      }
    },
    "/api/v1/kyc/vnin": {
      "post": {
        "description": "Submit a vNIN for identity verification. One per user across all businesses.",
        "operationId": "KycController_verifyVnin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyVninDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "vNIN verification submitted successfully."
          },
          "400": {
            "description": "Validation failed or the request body/params were malformed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired authentication.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource could not be found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "The resource already exists or is in a conflicting state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Verify virtual NIN (vNIN)",
        "tags": [
          "KYC Verification"
        ]
      }
    },
    "/api/v1/kyc/cac": {
      "post": {
        "description": "Submit a CAC RC number for business verification. One per business.",
        "operationId": "KycController_verifyCac",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyCacDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "CAC verification submitted successfully."
          },
          "400": {
            "description": "Validation failed or the request body/params were malformed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired authentication.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource could not be found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "The resource already exists or is in a conflicting state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Verify CAC",
        "tags": [
          "KYC Verification"
        ]
      }
    },
    "/api/v1/kyc/verifications/{reference}": {
      "get": {
        "description": "Fetches a verification record by its provider reference.",
        "operationId": "KycController_getVerification",
        "parameters": [
          {
            "name": "reference",
            "required": true,
            "in": "path",
            "description": "Provider reference ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Verification record retrieved successfully."
          },
          "401": {
            "description": "Missing, invalid, or expired authentication.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource could not be found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Get verification by reference",
        "tags": [
          "KYC Verification"
        ]
      }
    },
...
  },
  "info": {
    "title": "Dexxify API",
    "description": "Crypto Infrastructure API for Africa (Wallets, Payouts, Offramp, Onramp, KYC, KYB)",
    "version": "1.0",
    "contact": {
      "name": "Dexxify",
      "url": "https://www.dexxify.com",
      "email": "dexxifyhq@gmail.com"
    }
  },
...
  "servers": [
    {
      "url": "https://api.dexxify.com/api/v1",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "api-key": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "API key for /api/v1/* endpoints"
      }
    },
    "schemas": {
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "status": {
            "type": "number",
            "example": 400
          },
          "message": {
            "type": "string",
            "example": "Validation failed."
          },
          "errors": {
            "type": "object",
            "example": [
              "amount must not be less than 0"
            ]
          },
          "timestamp": {
            "type": "string",
            "example": "2026-09-17T12:00:00.000Z"
          }
        },
        "required": [
          "success",
          "status",
          "message",
          "timestamp"
        ]
      },
...
      "KycValidationDto": {
        "type": "object",
        "properties": {
          "first_name": {
            "type": "string",
            "description": "First name to match against record",
            "example": "John"
          },
          "last_name": {
            "type": "string",
            "description": "Last name to match against record",
            "example": "Doe"
          },
          "date_of_birth": {
            "type": "string",
            "description": "Date of birth (YYYY-MM-DD)",
            "example": "1990-01-15"
          },
          "selfie": {
            "type": "string",
            "description": "Base64-encoded selfie image for facial match",
            "example": "data:image/jpeg;base64,..."
          }
        }
      },
      "VerifyBvnDto": {
        "type": "object",
        "properties": {
          "bvn": {
            "type": "string",
            "description": "11-digit Bank Verification Number",
            "example": "12345678901"
          },
          "validation": {
            "description": "Optional fields to validate against the BVN record",
            "allOf": [
              {
                "$ref": "#/components/schemas/KycValidationDto"
              }
            ]
          }
        },
        "required": [
          "bvn"
        ]
      },
      "VerifyNinDto": {
        "type": "object",
        "properties": {
          "nin": {
            "type": "string",
            "description": "11-digit National Identification Number",
            "example": "12345678901"
          },
          "validation": {
            "description": "Optional fields to validate against the NIN record",
            "allOf": [
              {
                "$ref": "#/components/schemas/KycValidationDto"
              }
            ]
          }
        },
        "required": [
          "nin"
        ]
      },
      "VerifyVninDto": {
        "type": "object",
        "properties": {
          "vnin": {
            "type": "string",
            "description": "Virtual NIN (16-character alphanumeric)",
            "example": "AB1234567890CDEF"
          },
          "validation": {
            "description": "Optional fields to validate against the vNIN record",
            "allOf": [
              {
                "$ref": "#/components/schemas/KycValidationDto"
              }
            ]
          }
        },
        "required": [
          "vnin"
        ]
      },
      "VerifyCacDto": {
        "type": "object",
        "properties": {
          "rc_number": {
            "type": "string",
            "description": "CAC Registration Number",
            "example": "RC123456"
          },
          "registration_type": {
            "type": "string",
            "description": "Type of business registration",
            "enum": [
              "RC",
              "BN",
              "IT",
              "LP",
              "LLP"
            ],
            "example": "RC"
          },
          "registration_name": {
            "type": "string",
            "description": "Registered business name for additional matching",
            "example": "Acme Corp Ltd"
          }
        },
        "required": [
          "rc_number",
          "registration_type"
        ]
      },
...
    }
  }
}